+++ src/main/java/act/db/Dao.java
@@ -21,8 +21,6 @@ package act.db;
  */
 
 import act.Destroyable;
-import act.app.AppContextAware;
-import act.app.security.SecurityContextAware;
 
 import java.util.Collection;
 import java.util.List;
@@ -32,8 +30,7 @@ import java.util.List;
  * @param <ID_TYPE> the generic key type
  * @param <MODEL_TYPE> the generic model type
  */
-public interface Dao<ID_TYPE, MODEL_TYPE, QUERY_TYPE extends Dao.Query<MODEL_TYPE, QUERY_TYPE>>
-        extends AppContextAware, SecurityContextAware, Destroyable {
+public interface Dao<ID_TYPE, MODEL_TYPE, QUERY_TYPE extends Dao.Query<MODEL_TYPE, QUERY_TYPE>> extends Destroyable {
 
     /**
      * Returns the identifier type